# packages/contracts/openapi/guest-intel-service.json
{
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "ConciergeOut": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "escalated": {
            "title": "Escalated",
            "type": "boolean"
          },
          "escalation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Reason"
          },
          "handled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "sources": {
            "items": {},
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "id",
          "question",
          "answer",
          "sources",
          "outcome",
          "escalated",
          "created_at"
        ],
        "title": "ConciergeOut",
        "type": "object"
      },
      "GuestDnaOut": {
        "properties": {
          "average_sentiment": {
            "title": "Average Sentiment",
            "type": "number"
          },
          "complaint_themes": {
            "items": {
              "type": "string"
            },
            "title": "Complaint Themes",
            "type": "array"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "days_since_last_visit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Since Last Visit"
          },
          "favourite_items": {
            "items": {
              "type": "string"
            },
            "title": "Favourite Items",
            "type": "array"
          },
          "guest_id": {
            "format": "uuid",
            "title": "Guest Id",
            "type": "string"
          },
          "is_at_risk": {
            "title": "Is At Risk",
            "type": "boolean"
          },
          "preferences": {
            "items": {},
            "title": "Preferences",
            "type": "array"
          },
          "risk_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Risk Reasons",
            "type": "array"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          },
          "segment": {
            "title": "Segment",
            "type": "string"
          },
          "sentiment_label": {
            "title": "Sentiment Label",
            "type": "string"
          },
          "typical_gap_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Gap Days"
          }
        },
        "required": [
          "guest_id",
          "preferences",
          "favourite_items",
          "average_sentiment",
          "sentiment_label",
          "complaint_themes",
          "segment",
          "is_at_risk",
          "risk_score",
          "risk_reasons",
          "computed_at"
        ],
        "title": "GuestDnaOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PassageOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "category",
          "is_active"
        ],
        "title": "PassageOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Vesper Guest Intelligence",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/guest-intel/at-risk": {
      "get": {
        "description": "Guests drifting away, most at risk first.",
        "operationId": "at_risk_guest_intel_at_risk_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GuestDnaOut"
                  },
                  "title": "Response At Risk Guest Intel At Risk Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "At Risk",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/concierge/ask": {
      "post": {
        "description": "The guest chat. Rate limited per stay; answers only from the knowledge base.",
        "operationId": "guest_ask_guest_intel_concierge_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciergeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Guest Ask",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/concierge/escalations": {
      "get": {
        "description": "Questions the concierge could not answer \u2014 a staff queue and a content backlog.",
        "operationId": "escalations_guest_intel_concierge_escalations_get",
        "parameters": [
          {
            "in": "query",
            "name": "unhandled_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Unhandled Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConciergeOut"
                  },
                  "title": "Response Escalations Guest Intel Concierge Escalations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Escalations",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/concierge/escalations/{message_id}/handle": {
      "post": {
        "operationId": "handle_escalation_guest_intel_concierge_escalations__message_id__handle_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciergeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Handle Escalation",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/concierge/history": {
      "get": {
        "operationId": "guest_history_guest_intel_concierge_history_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConciergeOut"
                  },
                  "title": "Response Guest History Guest Intel Concierge History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Guest History",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/concierge/models": {
      "get": {
        "description": "What the configured Groq key can actually serve, for the settings page.",
        "operationId": "concierge_models_guest_intel_concierge_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Concierge Models Guest Intel Concierge Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Concierge Models",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/concierge/staff-ask": {
      "post": {
        "description": "The same concierge for staff \u2014 policy lookups without asking a manager.",
        "operationId": "staff_ask_guest_intel_concierge_staff_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciergeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Staff Ask",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/dna/{guest_id}": {
      "get": {
        "description": "The Guest DNA card: preference chips, sentiment and churn risk.",
        "operationId": "get_dna_guest_intel_dna__guest_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "guest_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Guest Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestDnaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dna",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/dna/{guest_id}/rebuild": {
      "post": {
        "operationId": "rebuild_dna_guest_intel_dna__guest_id__rebuild_post",
        "parameters": [
          {
            "in": "path",
            "name": "guest_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Guest Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestDnaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rebuild Dna",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/knowledge": {
      "get": {
        "operationId": "list_passages_guest_intel_knowledge_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PassageOut"
                  },
                  "title": "Response List Passages Guest Intel Knowledge Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Passages",
        "tags": [
          "guest-intel"
        ]
      },
      "post": {
        "description": "Add something the concierge may answer from, and reindex immediately.",
        "operationId": "create_passage_guest_intel_knowledge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Passage",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/knowledge/reindex": {
      "post": {
        "operationId": "reindex_guest_intel_knowledge_reindex_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reindex Guest Intel Knowledge Reindex Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reindex",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/offers": {
      "get": {
        "operationId": "list_offers_guest_intel_offers_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OfferOut"
                  },
                  "title": "Response List Offers Guest Intel Offers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Offers",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/offers/{offer_id}/approve": {
      "post": {
        "description": "The offer composer's Save. Nothing reaches a guest until this happens.",
        "operationId": "approve_offer_guest_intel_offers__offer_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferApprove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Offer",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/offers/{offer_id}/send": {
      "post": {
        "operationId": "send_offer_guest_intel_offers__offer_id__send_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Offer",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/sentiment": {
      "post": {
        "description": "Score a comment. Normally driven by the rating event, exposed for backfills.",
        "operationId": "score_sentiment_guest_intel_sentiment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentimentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Score Sentiment",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/sentiment/summary": {
      "get": {
        "operationId": "sentiment_summary_guest_intel_sentiment_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sentiment Summary",
        "tags": [
          "guest-intel"
        ]
      }
    },
    "/guest-intel/sentiment/trend": {
      "get": {
        "description": "Department sentiment over time \u2014 the trend chart.",
        "operationId": "sentiment_trend_guest_intel_sentiment_trend_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Sentiment Trend Guest Intel Sentiment Trend Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sentiment Trend",
        "tags": [
          "guest-intel"
        ]
      }
    },
...
  }
}